Refactor: Move metrics screens to separate package
This commit moves all UI components related to displaying metrics (PowerMetrics, EnvironmentMetrics, etc.) into a new `com.geeksville.mesh.ui.metrics` package.

This improves code organization by grouping related UI elements.
