+++ app/src/main/java/com/geeksville/mesh/ui/metrics/SignalMetrics.kt
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.ui.components
+package com.geeksville.mesh.ui.metrics
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.horizontalScroll
@@ -59,8 +59,12 @@ import com.geeksville.mesh.MeshProtos.MeshPacket
 import com.geeksville.mesh.R
 import com.geeksville.mesh.model.MetricsViewModel
 import com.geeksville.mesh.model.TimeFrame
-import com.geeksville.mesh.ui.components.CommonCharts.DATE_TIME_FORMAT
-import com.geeksville.mesh.ui.components.CommonCharts.MS_PER_SEC
+import com.geeksville.mesh.ui.metrics.CommonCharts.DATE_TIME_FORMAT
+import com.geeksville.mesh.ui.metrics.CommonCharts.MS_PER_SEC
+import com.geeksville.mesh.ui.components.LoraSignalIndicator
+import com.geeksville.mesh.ui.components.OptionLabel
+import com.geeksville.mesh.ui.components.SlidingSelector
+import com.geeksville.mesh.ui.components.SnrAndRssi
 import com.geeksville.mesh.util.GraphUtil.plotPoint
 
 @Suppress("MagicNumber")
